bmi_rdma: Remove return_active_nic_handle()
This function was already commented out, but remove it completely.

Since rdma_resolve_addr() handles binding a connection to an RDMA
device, this function is no longer needed. If for some reason in the
future we decide we do need it or we need to allow a user to specify
which device port to use, it will have to be made to work with librdmacm
and the changes made in commit 2fe7a0b88a19accc9e01b7d9650c420930753cff.