[Fix] bmi_rdma: Check if already disconnected in BMI_rdma_cancel()
After destroying an id, set its value back to NULL, and check that value
before trying to disconnect a connection in BMI_rdma_cancel().
Otherwise, if the timing is just right, an operation can get cancelled
while a server is trying to connect to another server that hasn't come
up yet. When that happens, it is possible BMI has already received a
connection error and disconnected by the time it receives the cancel
request from the job layer. If that is the case, we try to disconnect an
id that has already been destroyed and get a segfault. This issue can be
difficult to reproduce, but the best way to try to force it to happen is
to start the main server (listed first in the config file) after a
secondary server has already been started. Even then it isn't guaranteed
to happen, but it seems to be much more likely in that case.