Add experimental BMI InfiniBand module
This is essentially an exact copy of bmi_ib that will include future
experimental changes/fixes that have yet to be fully tested. Potentially
more elegant solutions are being considered that wouldn't require an
entirely separate module or would at least consolidate things, but for
now we are leaving that to future work.

As of now, to build this module use --with-experimental-ib when running
the configure script. This may change, but for now it requires bmi_ib to
be built as well, so --with-ib, --with-openib, or their variants must
also be used. However, naming changes are needed to make it compile
alongside bmi_ib. Subsequent commits will address that.