Build bmi_ib OR bmi_ib_exp, not both
The --with-ib, --with-openib, or variant options are still required when
using --with-experimental-ib, but if the latter option is specified
don't try to build both bmi_ib and bmi_ib_exp. Since bmi_ib_exp is a
copy of bmi_ib, building both would require changing the names of most
variables and functions in bmi_ib_exp to avoid collisions. Since the end
goal is for the experimental changes to eventually make their way into
bmi_ib or for bmi_ib_exp to completely replace bmi_ib in the future, it
doesn't make sense to change names. They would need to be changed back
in the future, and this would potentially make the process of moving a
change from "experimental" to "production" more tedious and error-prone.
Instead, make the modules mutually exclusive, and build only bmi_ib_exp
when --with-experimental-ib is specified.