Fix segfault issue when shutting down restarted servers
When servers are shutdown and restarted without wiping and
re-initializing the storage, a different path is followed upon startup
than when first starting servers with freshly initialized storage. In
the latter case, msgpairs are exchanged between the servers right way,
so new connections are setup and established as soon as the servers are
up. But in the former case, this msgpair exchange does not occur, so
connections are not setup until triggered by another event (e.g. when a
client connects). So if a server is shutdown before that event ever
occurs, the protection domain, completion channel, and completion queue
will have never been created. Therefore, trying to clean them up would
result in a segfault. Fix this by checking for their existence before
attempting to clean them up.