The lookup is weird because there is a need to simply "lookup" an entry in a specific dirdata obj - but the actual lookup takes a path you are looking up Starting in some directory, possibly spanning mutiple servers.  Apparently At some point someone hijacked the management routine "mgmt_get_dirent" which did the right thing.  Later someone modified it so that it sort of looked like the sys-lookup call - rendering it broken. This leads to a philosophical question "should we be using a mgmt routine in normal operations?" (note that mgmt tools can use regular routines).  Part of the question comes down to the fact that this routine is never called from the client, like all other system interface routines, but only server to server.
So, in these changes I rewrote the mgmt routine back to its simplied function,
which is what we need, but left it mgmt for now as changing that is messy.
We need a study of the mgmt interface to see if we still need it, and what
it should look like.  Most of it has not been looked at since V3 started and
the relevant code has not been converted to V3.

"client" calls to msgpairarray that use a comp_fn AND access the parent
frame (MPA pushes a frame) have an extra step to retrieve the parent smcb from the
current smcb.  Some of the code has not been updated.  I fixed lookup.  There
are others that need this fix.